Attempt to display HUD points

+++ source/mesh_builder/mesh_builder_widget.h
@@ -62,10 +62,10 @@ private:
     std::unique_ptr<QOpenGLTexture> photo_texture;
     int photo_width;
     int photo_height;
-    float photo_x_low;
-    float photo_x_high;
-    float photo_y_low;
-    float photo_y_high;
+    int photo_x_low;
+    int photo_x_high;
+    int photo_y_low;
+    int photo_y_high;
     float photo_aspect;
 
     QVector3D viewer_pos;
